Support flake8 version 3
With version 3 of flake8 it removed the `config_options` attribute and changed
it to a custom keyword argument. This is providing support for both variants
by first trying the call with the keyword argument and then, if it fails, by
removing that keyword argument and manually adding the option to
`config_options`.